fixing race condition in evolve_async where it now clears the noise model before returning promise
